requirements.txt
Django===2.1.3
psycopg2-binary===2.7.5
argon2-cffi===18.1.0
cffi===1.11.5
passlib===1.7.1
djangorestframework===3.9.0
markdown===3.0.1
django-filter===2.0.0